web_time/requirements.txt
Django==3.2.12
djangorestframework==3.14.0
Pillow
psycopg2-binary
django-ckeditor==6.0.0

